data_pipeline/sentiment/requirements.txt
pymongo
newsapi-python
dotenv-python
apache-airflow-providers-mongo==3.0.0